backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain and LangGraph
langchain==0.1.16
langchain-community==0.0.34 
langchain-google-genai==0.0.6
langchain-groq==0.0.1
langgraph==0.0.20


# Vector Database
qdrant-client==1.7.0

# Code Processing
GitPython==3.1.40
tree-sitter==0.20.4
tree-sitter-python==0.23.6

# Embeddings and AI
google-generativeai==0.3.2
groq==0.4.1
sentence-transformers==2.2.2

# Utilities
httpx==0.25.2
aiofiles==23.2.1
loguru==0.7.2
tenacity==8.2.3
tiktoken==0.5.2  # Still useful for token counting

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0


python-dotenv